Added tne axios call to the frontend and also made the backend more modular and clean to read
