+++ server/__init__.py
@@ -1,4 +1,5 @@
 # server/__init__.py
 
-from .server import app, duration_detector, parse_srt, srt_time_to_seconds
 import pdf_generator
+import routes
+import utils
