+++ server/routes/__init__.py
@@ -1,0 +1,3 @@
+from utils import duration_detector, parse_srt
+
+from . import pdf_routes, transcribe_routes
