+++ transcript-client/src/hooks/useAxios.ts
@@ -1,0 +1,7 @@
+import { AxiosPrivateClient } from "src/utils/axios"
+
+const useAxios = () => {
+    return AxiosPrivateClient
+}
+
+export default useAxios
