+++ transcript-client/src/utils/axios.ts
@@ -1,0 +1,12 @@
+import axios from "axios"
+
+const API_URL = process.env.PROD ? "/api" : "http://localhost:8000"
+
+export const AxiosPrivateClient = axios.create({
+  baseURL: API_URL,
+  // headers: {
+  //   "Content-Type": "application/json"
+  // },
+})
+
+
